add gameplay statistics page
