+++ package.json
@@ -24,6 +24,7 @@
     "posthog-js": "^1.96.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.2",
     "typescript": "^5.1.3"
   },
   "scripts": {
