+++ src/components/index.ts
@@ -4,3 +4,5 @@ export * from "./modals";
 export * from "./Header";
 export * from "./CountrySpecsList";
 export * from "./CountryArea";
+export * from "./PlayGame";
+export * from "./App";
